requirements.txt
astroid==2.4.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
psycopg2==2.9.1
pylint==2.6.0
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.4.22
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
